pyproject.toml
[project]
name = "leads-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-core>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.9",
    "openai>=2.24.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "pytz>=2026.1.post1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
